MongoDB/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-prometheus==2.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.2
mongoengine==0.21.0
prometheus-client==0.9.0
pymongo==3.11.1
pytz==2020.4
sqlparse==0.4.1
